+++ packages/cli/core/src/lib/flags/snapshotCommonFlags.ts
@@ -28,15 +28,6 @@ export const previewLevel = () => ({
   }),
 });
 
-export const organization = (description: string) => ({
-  organization: Flags.string({
-    char: 'o',
-    helpValue: 'targetorganizationg7dg3gd',
-    required: false,
-    description: `${description} If not specified, the organization you are connected to will be used.`,
-  }),
-});
-
 export const snapshotId = () => ({
   snapshotId: Flags.string({
     char: 's',
