+++ packages/cli/core/src/lib/utils/platform.ts
@@ -1,0 +1,9 @@
+import {Config} from '@coveo/cli-commons/config/config';
+
+export function getTargetOrg(config: Config, target?: string) {
+  if (target) {
+    return target;
+  }
+  const cfg = config.get();
+  return cfg.organization;
+}
